requirements.txt
pytest==7.4.0
selene==2.0.0rc3.post3
allure-pytest